% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspeciesaccumulation-package.R
\docType{package}
\name{ggspeciesaccumulation-package}
\alias{ggspeciesaccumulation}
\alias{ggspeciesaccumulation-package}
\title{ggspeciesaccumulation: Species accumulation curves on tidy data}
\description{
A ggplot2 stat the produces simple species accumulation curves.
}
\author{
\strong{Maintainer}: Sam Herniman \email{sam@pandionlabs.dev} (\href{https://orcid.org/0000-0001-7663-2897}{ORCID})

Authors:
\itemize{
  \item Sam Herniman \email{sam@pandionlabs.dev} (\href{https://orcid.org/0000-0001-7663-2897}{ORCID})
}

}
\keyword{internal}
